13 Sept 2024, 20:58

+++ Datatypes/Arrays/Questions/findLargest.cpp
@@ -15,7 +15,7 @@ int findMax(std::vector<T>& arr)
         }
     
     }
-    return static_cast<int>(max);
+    return max;
 }
 
 int main()
